Корректировка методов

+++ shop/src/main/java/org/vicrul/shop/dao/CustomerDAO.java
@@ -11,7 +11,7 @@ public interface CustomerDAO {
 	List<Customer> findBySurname(String surname);
 	List<Customer> findWhoBoughtThisProduct(String title, long minTimes);
 	List<Customer> intervalCost(long minPrice, long maxPrice);
-	List<Customer> findPassiveCustomers(int count);
+	List<Customer> findPassiveCustomers(long count);
 	void createView(String startDate, String endDate);
 	void dropView();
 	List<Customer> getCustomersInPeriod();
